+++ src/types/customIds.ts
@@ -1,3 +1,4 @@
 export const PRIME_MODAL = 'botluck:prime-modal';
 export const PRIME_MODAL_TEMPLATE = 'template';
 export const PRIME_MODAL_THEME = 'theme';
+export const PRIME_MODAL_BLIND = 'blind';